moved functionality to a shared location
